interface_impl: Add common CredentialIssuerCertificate interface impl
Extract common `Aliro::Interface::CredentialIssuerCertificate`
implementation
from applications.
