+++ applications/aliro-access-control-app/src/aliro/Kconfig
@@ -44,7 +44,6 @@ config DOOR_LOCK_CREDENTIAL_ISSUER_CA
     bool "Support for Credential Issuer CA public key"
     default y if DOOR_LOCK_STEP_UP_PHASE
     depends on DOOR_LOCK_STEP_UP_PHASE
-    select MBEDTLS_X509_LIBRARY
     help
         If enabled, the Credential Issuer CA public key is supported.
         The Credential Issuer CA public key is used to verify the signature of the Credential Issuer certificate.
