+++ applications/matter-aliro-door-lock-app/Kconfig.defconfig
@@ -102,6 +102,12 @@ configdefault DOOR_LOCK_ALIRO_INTERFACE_IMPL_READER
 configdefault DOOR_LOCK_ALIRO_INTERFACE_IMPL_CRYPTO
 	default y
 
+configdefault DOOR_LOCK_ALIRO_INTERFACE_IMPL_CI_CERT
+	default y
+
+configdefault DOOR_LOCK_ALIRO_INTERFACE_IMPL_CI_CERT_VALIDATION
+	default y if DOOR_LOCK_CREDENTIAL_ISSUER_CA
+
 configdefault DOOR_LOCK_ALIRO_INTERFACE_IMPL_OS_BLE_STACK_TIMERS
 	default DOOR_LOCK_BLE_UWB_MAX_SESSIONS if DOOR_LOCK_BLE_UWB
