+++ subsys/aliro/interface_impl/CMakeLists.txt
@@ -9,3 +9,4 @@ add_subdirectory_ifdef(CONFIG_DOOR_LOCK_ALIRO_INTERFACE_IMPL_LOG log)
 add_subdirectory_ifdef(CONFIG_DOOR_LOCK_ALIRO_INTERFACE_IMPL_UWB uwb)
 add_subdirectory_ifdef(CONFIG_DOOR_LOCK_ALIRO_INTERFACE_IMPL_READER reader)
 add_subdirectory_ifdef(CONFIG_DOOR_LOCK_ALIRO_INTERFACE_IMPL_CRYPTO crypto)
+add_subdirectory_ifdef(CONFIG_DOOR_LOCK_ALIRO_INTERFACE_IMPL_CI_CERT ci_cert)
